requirements.txt
Django
pillow
django-cleanup
django-allauth
django-htmx
daphne
shortuuid
channels